requirements.txt
tqdm
requests
psycopg2-binary
